requirements.txt
asgiref==3.2.7
autopep8==1.5.3
Cython==0.29.21
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
django-sendgrid-v5==0.8.1
djangorestframework==3.11.0
fpdf==1.7.2
future==0.18.2
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.6.0
Pygments==2.6.1
pytesseract==0.3.4
python-decouple==3.3
python-http-client==3.2.7
python-magic==0.4.18
pytz==2020.1
sendgrid==6.3.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.1.0
